PP-4131 Move charge calculator classes to charge.util package
with @DanailMinchev
