+++ src/main/java/uk/gov/pay/connector/charge/util/RefundCalculator.java
@@ -1,6 +1,7 @@
-package uk.gov.pay.connector.util.charge;
+package uk.gov.pay.connector.charge.util;
 
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
+import uk.gov.pay.connector.common.service.charge.CorporateCardSurchargeCalculator;
 import uk.gov.pay.connector.refund.model.domain.RefundEntity;
 import uk.gov.pay.connector.refund.model.domain.RefundStatus;
