+++ src/test/java/uk/gov/pay/connector/charge/util/RefundCalculatorTest.java
@@ -1,0 +1,5 @@
+import static org.junit.Assert.*;
+
+public class RefundCalculatorTest {
+
+}
